rename noralize_cell to normalize_cell get rid of unnecessary print
