+++ command_helper/get_documentation.py
@@ -3,13 +3,13 @@ from os import path
 try:
     from parser_utils.file_formatter import rf_table_name
     from parser_utils.util import get_index_name, normalise_path
-    from noralize_cell import get_data_from_json
+    from normalize_cell import get_data_from_json
     from db_json_settings import DBJsonSetting
     from utils.util import kw_equals_kw_candite
 except:
     from ..dataparser.parser_utils.file_formatter import rf_table_name
     from ..dataparser.parser_utils.util import get_index_name, normalise_path
-    from ..command_helper.noralize_cell import get_data_from_json
+    from ..command_helper.normalize_cell import get_data_from_json
     from ..setting.db_json_settings import DBJsonSetting
     from ..command_helper.utils.util import kw_equals_kw_candite
