+++ command_helper/jump_to_file.py
@@ -3,10 +3,10 @@ import re
 
 try:
     from parser_utils.file_formatter import lib_table_name
-    from noralize_cell import get_data_from_json
+    from normalize_cell import get_data_from_json
 except ImportError:
     from ..dataparser.parser_utils.file_formatter import lib_table_name
-    from .noralize_cell import get_data_from_json
+    from .normalize_cell import get_data_from_json
 
 
 class JumpToFile(object):
