+++ command_helper/workspace_objects.py
@@ -1,9 +1,9 @@
 from os import listdir, path
 try:
-    from .noralize_cell import get_data_from_json
+    from .normalize_cell import get_data_from_json
     from db_json_settings import DBJsonSetting
 except:
-    from ..command_helper.noralize_cell import get_data_from_json
+    from ..command_helper.normalize_cell import get_data_from_json
     from ..setting.db_json_settings import DBJsonSetting
