+++ commands/show_documentation.py
@@ -4,7 +4,7 @@ from ..setting.setting import get_setting
 from ..setting.setting import SettingObject
 from .query_completions import get_index_file
 from ..command_helper.utils.get_text import get_line
-from ..command_helper.noralize_cell import ReturnKeywordAndObject
+from ..command_helper.normalize_cell import ReturnKeywordAndObject
 from ..command_helper.get_metadata import get_rf_table_separator
 from ..command_helper.get_documentation import GetKeywordDocumentation
